+++ src/main/kotlin/Application.kt
@@ -1,5 +1,3 @@
-import domain.CarRacingGameController
-
 fun main() {
     val carRacingGameController = CarRacingGameController()
