+++ src/main/kotlin/domain/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package data.generator
+package domain
 
 interface NumberGenerator {
     fun generate(): Int
